881 - make consistent use of block level definition (#882)
* Bump the (quite old) scalatest version and add scalacheck support

 * this would allow us to introduce property-based testing for the
   fork handling code

* Refactor to block level as 64-bit int value, fix all tests

 * correct all database types for level as bigint
 * correct all code modeling levels to use the new alias
 * give an explicit class to BlockReference for better matching
 * simplify the node operator tests and add explanatory comments
 * add test case for loading blocks not indexed in db

* Resolve some leftovers

+++ conseil-lorre/src/main/scala/tech/cryptonomic/conseil/indexer/logging/LorreProgressLogging.scala
@@ -16,7 +16,7 @@ trait LorreProgressLogging {
     * @param processStartNanos a nano-time from jvm monotonic time, used to identify when the whole processing operation began
     * @param processed how many entities were processed at the current checkpoint
     */
-  def logProcessingProgress(entityName: String, totalToProcess: Int, processStartNanos: Long)(
+  def logProcessingProgress(entityName: String, totalToProcess: Long, processStartNanos: Long)(
       processed: Int
   ): Unit = {
     val elapsed = System.nanoTime() - processStartNanos
